rename editor.js
